+++ noodles-bam/src/bai/index.rs
@@ -83,7 +83,9 @@ impl Index {
     }
 }
 
-impl BinningIndex<ReferenceSequence> for Index {
+impl BinningIndex for Index {
+    type ReferenceSequence = ReferenceSequence;
+
     /// Returns a list of indexed reference sequences.
     ///
     /// This list is parallel to the reference sequences defined in the associated BAM file.
@@ -96,7 +98,7 @@ impl BinningIndex<ReferenceSequence> for Index {
     /// let index = bai::Index::default();
     /// assert!(index.reference_sequences().is_empty());
     /// ```
-    fn reference_sequences(&self) -> &[ReferenceSequence] {
+    fn reference_sequences(&self) -> &[Self::ReferenceSequence] {
         &self.reference_sequences
     }