+++ noodles-tabix/src/index.rs
@@ -115,7 +115,9 @@ impl Index {
     }
 }
 
-impl BinningIndex<ReferenceSequence> for Index {
+impl BinningIndex for Index {
+    type ReferenceSequence = ReferenceSequence;
+
     /// Returns a list of indexed reference sequences.
     ///
     /// # Examples
@@ -126,7 +128,7 @@ impl BinningIndex<ReferenceSequence> for Index {
     /// let index = tabix::Index::default();
     /// assert!(index.reference_sequences().is_empty());
     /// ```
-    fn reference_sequences(&self) -> &[ReferenceSequence] {
+    fn reference_sequences(&self) -> &[Self::ReferenceSequence] {
         &self.reference_sequences
     }